+++ market/src/main/java/university/market/chat/room/domain/ChatVO.java
@@ -1,7 +1,6 @@
-package university.market.chat.domain;
+package university.market.chat.room.domain;
 
 import java.sql.Timestamp;
-import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
