+++ market/src/main/resources/mapper/chatMember-mapper.xml
@@ -3,9 +3,9 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="university.market.chat.mapper.ChatMemberMapper">
+<mapper namespace="university.market.chat.room.mapper.ChatMemberMapper">
 
-    <resultMap id="ChatMemberResultMap" type="university.market.chat.domain.ChatMemberVO">
+    <resultMap id="ChatMemberResultMap" type="university.market.chat.room.domain.ChatMemberVO">
         <id property="id" column="id"/>
         <result property="chatAuth" column="chat_auth"/>
         <result property="createdAt" column="created_at"/>
@@ -17,7 +17,7 @@
             <result property="university" column="member_university"/>
         </association>
 
-        <association property="chat" javaType="university.market.chat.domain.ChatVO">
+        <association property="chat" javaType="university.market.chat.room.domain.ChatVO">
             <id property="id" column="chat_id"/>
             <result property="title" column="chat_title"/>
             <result property="createdAt" column="chat_created_at"/>
